refactor(CvCard): implemented new CvCard component in DigitalCV, added isNotNil ramda util, tweaked styles
