+++ src/components/CvCard/CvCard.tsx
@@ -3,6 +3,7 @@ import { View } from 'react-native'
 
 import { Colors } from '../../styles'
 import { WithChildren } from '../../types/react.types'
+import { isNotNil } from '../../utils/ramda'
 import ListCardHeader from '../ListCardHeader'
 import Optional from '../Optional'
 import Text, { TextAlign } from '../Typography'
@@ -22,13 +23,15 @@ const CvCard = ({ title, fallback, count, badgeColor = Colors.white, onEdit, chi
       color={badgeColor}
       count={count || 0}
       header={title}
-      hasCountBadge={!!count}
+      hasCountBadge={isNotNil(count) as boolean}
       rightAccessory={<EditButton onPress={onEdit} />}
       hasBorder
     />
-    <Optional condition={!!children} fallback={<Text.Body align={TextAlign.center}>{fallback}</Text.Body>}>
-      {children}
-    </Optional>
+    <View style={styles.content}>
+      <Optional condition={!!children} fallback={<Text.Body align={TextAlign.center}>{fallback}</Text.Body>}>
+        {children}
+      </Optional>
+    </View>
   </View>
 )
