+++ src/components/CvCard/EditButton.styles.ts
@@ -1,0 +1,12 @@
+import { StyleSheet, ViewStyle } from 'react-native'
+import { colors, Colors } from 'styles'
+
+const styles = {
+  container: {
+    elevation: 3,
+    backgroundColor: colors[Colors.white],
+    borderRadius: 15,
+  } as ViewStyle,
+}
+
+export default StyleSheet.create(styles)
