+++ src/components/CvCard/EditButton.tsx
@@ -1,18 +1,17 @@
 import React from 'react'
-import { Pressable } from 'react-native'
+import { Pressable, View } from 'react-native'
 
 import { EditIcon } from '../../assets/images'
-import { Colors, colors } from '../../styles'
+import styles from './EditButton.styles'
 
 interface EditButtonProps {
   onPress: () => void
 }
 const EditButton = ({ onPress }: EditButtonProps) => (
-  <Pressable
-    onPress={onPress}
-    style={({ pressed }) => [{ backgroundColor: colors[Colors.backgroundGrey], opacity: pressed ? 0.5 : 1 }]}
-  >
-    <EditIcon />
+  <Pressable onPress={onPress} hitSlop={20} style={({ pressed }) => [{ opacity: pressed ? 0.5 : 1 }]}>
+    <View style={styles.container}>
+      <EditIcon />
+    </View>
   </Pressable>
 )
