+++ src/components/CvCard/index.ts
@@ -1,5 +1,3 @@
 import CvCard from './CvCard'
-import CvCardOld from './CvCardOld'
 
 export default CvCard
-export { CvCardOld }
