+++ src/modules/DigitalCv/DigitalCv.tsx
@@ -1,6 +1,6 @@
 import { StackNavigationProp } from '@react-navigation/stack'
 import { FirstTimeCard, ViewContainer } from 'components'
-import CvCard, { CvCardOld } from 'components/CvCard'
+import CvCard from 'components/CvCard'
 import HomeHeader from 'components/HomeHeader'
 import { HomeNavigationRoutes, HomeNavigatorParamsList } from 'modules/HomeNavigation/HomeNavigation.types'
 import React from 'react'
@@ -30,46 +30,39 @@ const DigitalCv = ({ navigation, biography }: Props) => {
         >
           <Text.Body align={TextAlign.center}>{biography}</Text.Body>
         </CvCard>
-        <CvCardOld
-          cardTitle={t('About')}
-          defaultText={t('Your biography is one of the first things recruiters look at. Write a great one!')}
-          content={biography}
-          hasCountBadge={false}
-          onEdit={() => navigation.navigate(HomeNavigationRoutes.About)}
-        />
-        <CvCardOld
+        <CvCard
           count={0}
           badgeColor={Colors.primaryBlue}
-          cardTitle={t('Experience')}
-          defaultText={t('Where do you currently work?')}
+          title={t('Experience')}
+          fallback={t('Where do you currently work?')}
           onEdit={() => navigation.navigate(HomeNavigationRoutes.Experience)}
         />
-        <CvCardOld
+        <CvCard
           count={0}
           badgeColor={Colors.primaryRed}
-          cardTitle={t('Education')}
-          defaultText={t('Which school, university or college did you attend?')}
+          title={t('Education')}
+          fallback={t('Which school, university or college did you attend?')}
           onEdit={() => navigation.navigate(HomeNavigationRoutes.Education)}
         />
-        <CvCardOld
+        <CvCard
           count={0}
           badgeColor={Colors.primaryGreen}
-          cardTitle={t('My skills')}
-          defaultText={t('Tell us what you are great at.')}
+          title={t('My skills')}
+          fallback={t('Tell us what you are great at.')}
           onEdit={() => navigation.navigate(HomeNavigationRoutes.Skills)}
         />
-        <CvCardOld
+        <CvCard
           count={0}
           badgeColor={Colors.primaryYellow}
-          cardTitle={t('Completed courses')}
-          defaultText={t('Have you completed any courses yet?')}
+          title={t('Completed courses')}
+          fallback={t('Have you completed any courses yet?')}
           onEdit={() => navigation.navigate(HomeNavigationRoutes.MyCourses)}
         />
-        <CvCardOld
+        <CvCard
           count={0}
           badgeColor={Colors.secondaryPurple}
-          cardTitle={t('Completed challenges')}
-          defaultText={t('Have you completed any challenges yet?')}
+          title={t('Completed challenges')}
+          fallback={t('Have you completed any challenges yet?')}
           onEdit={() => navigation.navigate(HomeNavigationRoutes.MyChallenges)}
         />
       </ScrollView>
