+++ src/utils/ramda.ts
@@ -1,0 +1,3 @@
+import { complement, isNil } from 'ramda'
+
+export const isNotNil = complement(isNil)
